refactor(repl): migrate user-design commands into UserDesignCommandSet
Move synthesis, place_and_route, gen_bitStream_binary, run_FABulous_bitstream
and gen_user_design_wrapper into UserDesignCommandSet, and fold compile_design
and run_simulation in as methods of the same set, dropping the standalone
cmd_compile_design/cmd_run_simulation modules and the free-function
'self: FABulousREPL' pattern. The shell class is now a thin host that only
registers the CommandSets plus exit/quit/q and the run_script builtin
override. Update the compile/simulation tests to patch cmd_user_design, and
finish the CLI->REPL wording in module docstrings.

stack-info: PR: https://github.com/FPGA-Research/FABulous/pull/894, branch: KelvinChung2000/stack/8
